pyproject.toml
[project]
name = "poc-dta-mcp"
version = "0.1.0"
description = "POC: Data Analytics MCP Server with PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
